completed simon says

+++ keil/helpers.h
@@ -4,7 +4,7 @@
 #include "driverlib.h"
 
 unsigned char reverse(unsigned char b);
-double r2(void);
+float r2(void);
 void delayTimer(uint8_t seconds);
 
 #endif
